Split files to components
